# added performance boost to convert case