--- a/lib/JSONAPISerializer.js
+++ b/lib/JSONAPISerializer.js
@@ -13,6 +13,13 @@ const {
 
 const { validateOptions, validateDynamicTypeOptions, validateError } = require('./validator');
 
+// Cache of strings to convert to their converted values per conversion type
+const convertCaseMap = {
+  camelCase: {},
+  kebabCase: {},
+  snakeCase: {}
+};
+
 /**
  * JSONAPISerializer class.
  *
@@ -794,13 +801,25 @@ module.exports = class JSONAPISerializer {
 
         switch (convertCaseOptions) {
           case 'snake_case':
-            converted = toSnakeCase(data);
+            converted = convertCaseMap.snakeCase[data];
+            if (!converted) {
+              converted = toSnakeCase(data);
+              convertCaseMap.snakeCase[data] = converted;
+            }
             break;
           case 'kebab-case':
-            converted = toKebabCase(data);
+            converted = convertCaseMap.kebabCase[data];
+            if (!converted) {
+              converted = toKebabCase(data);
+              convertCaseMap.kebabCase[data] = converted;
+            }
             break;
           case 'camelCase':
-            converted = toCamelCase(data);
+            converted = convertCaseMap.camelCase[data];
+            if (!converted) {
+              converted = toCamelCase(data);
+              convertCaseMap.camelCase[data] = converted;
+            }
             break;
           default: // Do nothing
         }