--- a/lib/JSONAPISerializer.js
+++ b/lib/JSONAPISerializer.js
@@ -8,17 +8,11 @@ const {
   set,
   toCamelCase,
   toKebabCase,
-  toSnakeCase
+  toSnakeCase,
+  LRU
 } = require('./helpers');
 
 const { validateOptions, validateDynamicTypeOptions, validateError } = require('./validator');
-
-// Cache of strings to convert to their converted values per conversion type
-const convertCaseMap = {
-  camelCase: {},
-  kebabCase: {},
-  snakeCase: {}
-};
 
 /**
  * JSONAPISerializer class.
@@ -31,11 +25,19 @@ const convertCaseMap = {
  *
  * @class JSONAPISerializer
  * @param {object} [opts] Global options.
+ * @param {integer} [convertCaseCacheSize=5000] Size of cache used for convertCase, 0 results in an infinitely sized cache
  */
 module.exports = class JSONAPISerializer {
-  constructor(opts) {
+  constructor(opts, convertCaseCacheSize = 5000) {
     this.opts = opts || {};
     this.schemas = {};
+
+    // Cache of strings to convert to their converted values per conversion type
+    this.convertCaseMap = {
+      camelCase: new LRU(convertCaseCacheSize),
+      kebabCase: new LRU(convertCaseCacheSize),
+      snakeCase: new LRU(convertCaseCacheSize)
+    };
   }
 
   /**
@@ -801,24 +803,24 @@ module.exports = class JSONAPISerializer {
 
         switch (convertCaseOptions) {
           case 'snake_case':
-            converted = convertCaseMap.snakeCase[data];
+            converted = this.convertCaseMap.snakeCase.get(data);
             if (!converted) {
               converted = toSnakeCase(data);
-              convertCaseMap.snakeCase[data] = converted;
+              this.convertCaseMap.snakeCase.set(data, converted);
             }
             break;
           case 'kebab-case':
-            converted = convertCaseMap.kebabCase[data];
+            converted = this.convertCaseMap.kebabCase.get(data);
             if (!converted) {
               converted = toKebabCase(data);
-              convertCaseMap.kebabCase[data] = converted;
+              this.convertCaseMap.kebabCase.set(data, converted);
             }
             break;
           case 'camelCase':
-            converted = convertCaseMap.camelCase[data];
+            converted = this.convertCaseMap.camelCase.get(data);
             if (!converted) {
               converted = toCamelCase(data);
-              convertCaseMap.camelCase[data] = converted;
+              this.convertCaseMap.camelCase.set(data, converted);
             }
             break;
           default: // Do nothing