--- a/lib/JSONAPISerializer.js
+++ b/lib/JSONAPISerializer.js
@@ -25,7 +25,7 @@ const { validateOptions, validateDynamicTypeOptions, validateError } = require('
  *
  * @class JSONAPISerializer
  * @param {object} [opts] Global options.
- * @param {integer} [convertCaseCacheSize=5000] Size of cache used for convertCase, 0 results in an infinitely sized cache
+ * @param {number} [convertCaseCacheSize=5000] Size of cache used for convertCase, 0 results in an infinitely sized cache
  */
 module.exports = class JSONAPISerializer {
   constructor(opts, convertCaseCacheSize = 5000) {