--- a/lib/helpers.js
+++ b/lib/helpers.js
@@ -9,6 +9,7 @@ const {
   toCamelCase
 } = require('30-seconds-of-code');
 const set = require('lodash.set');
+const LRU = require('./lru-cache');
 
 // https://github.com/you-dont-need/You-Dont-Need-Lodash-Underscore#_get
 const get = (obj, path, defaultValue) =>
@@ -27,5 +28,6 @@ module.exports = {
   transform,
   toKebabCase,
   toSnakeCase,
-  toCamelCase
+  toCamelCase,
+  LRU
 };