I ran the benchmark before and after the changes. It seems to have made deserialization MUCH faster, and a smaller performance boost to serialize.
Resolves #84

# Before convertCase optimization

Platform info:
==============
Darwin 18.6.0 x64
Node.JS: 10.15.1
V8: 6.8.275.32-node.12
Intel(R) Core(TM) i5-5287U CPU @ 2.90GHz × 4

Suite:
==============
serializeAsync x 71,616 ops/sec ±1.76% (75 runs sampled)
serialize x 110,926 ops/sec ±4.38% (81 runs sampled)
serializeConvertCase x 70,255 ops/sec ±2.48% (91 runs sampled)
deserializeAsync x 152,227 ops/sec ±2.29% (80 runs sampled)
deserialize x 394,181 ops/sec ±1.24% (91 runs sampled)
deserializeConvertCase x 30,294 ops/sec ±1.01% (92 runs sampled)
serializeError x 284,856 ops/sec ±1.06% (92 runs sampled)
serializeError with a JSON API error object x 12,753,945 ops/sec ±4.90% (91 runs sampled)



# After convertCase optimization

Platform info:
==============
Darwin 18.6.0 x64
Node.JS: 10.15.1
V8: 6.8.275.32-node.12
Intel(R) Core(TM) i5-5287U CPU @ 2.90GHz × 4

Suite:
==============
serializeAsync x 70,198 ops/sec ±3.08% (76 runs sampled)
serialize x 119,645 ops/sec ±3.64% (85 runs sampled)
serializeConvertCase x 91,489 ops/sec ±3.21% (88 runs sampled)
deserializeAsync x 144,878 ops/sec ±5.55% (70 runs sampled)
deserialize x 361,896 ops/sec ±3.33% (87 runs sampled)
deserializeConvertCase x 133,200 ops/sec ±1.91% (91 runs sampled)
serializeError x 281,794 ops/sec ±1.05% (84 runs sampled)
serializeError with a JSON API error object x 12,884,729 ops/sec ±1.89% (86 runs sampled)