@alechirsch 2019-08-21
Added LRU cache, here are the new benchmarks

```
serializeAsync x 62,434 ops/sec ±13.04% (69 runs sampled)
serialize x 125,513 ops/sec ±2.49% (81 runs sampled)
serializeConvertCase x 91,754 ops/sec ±2.83% (88 runs sampled)
deserializeAsync x 161,124 ops/sec ±3.41% (80 runs sampled)
deserialize x 392,788 ops/sec ±1.19% (89 runs sampled)
deserializeConvertCase x 114,102 ops/sec ±0.55% (88 runs sampled)
serializeError x 267,726 ops/sec ±3.77% (83 runs sampled)
serializeError with a JSON API error object x 12,156,898 ops/sec ±3.78% (85 runs sampled)
```